test: move gotk network tests to integration, use flux-operator in workflow test
gotk bootstrap tests download manifests.tar.gz from GitHub at runtime.
When GitHub is unreachable from the CI runner the unit tests either fail
(connection refused) or cause the test binary to timeout (slow network
hang consumes the 30s budget shared with other tests in the package).

- TestGenerateGotkBootstrap / TestGenerateGotkBootstrapWithOptions moved
  to bootstrap_generator_integration_test.go with //go:build integration
- TestWorkflowInterface uses flux-operator mode (no network required)
  instead of gotk, which was testing the interface not the mode
